+++ packages/nimble-components/src/anchor/tests/anchor.spec.ts
@@ -31,12 +31,12 @@ describe('Anchor', () => {
 
     it('should set the "control" class on the internal control', async () => {
         await connect();
-        expect(element.control.classList.contains('control')).toBe(true);
+        expect(element.control!.classList.contains('control')).toBe(true);
     });
 
     it('should set the `part` attribute to "control" on the internal control', async () => {
         await connect();
-        expect(element.control.part.contains('control')).toBe(true);
+        expect(element.control!.part.contains('control')).toBe(true);
     });
 
     const attributeNames: { name: string }[] = [
@@ -85,7 +85,7 @@ describe('Anchor', () => {
                 element.setAttribute(attribute.name, 'foo');
                 await waitForUpdatesAsync();
 
-                expect(element.control.getAttribute(attribute.name)).toBe(
+                expect(element.control!.getAttribute(attribute.name)).toBe(
                     'foo'
                 );
             });