+++ packages/nimble-components/src/table-column/base/types.ts
@@ -15,15 +15,12 @@ export interface TableCellState<
     columnConfig: TColumnConfig;
 }
 
-export interface TableColumnWithPlaceholderColumnConfig {
-    placeholder: string;
-}
-
 /**
  * The event details for the 'delegated-event' CustomEvent
  */
 export interface DelegatedEventEventDetails {
     originalEvent: Event;
+    recordId: string;
 }
 
 /**